archive_20211013/requirements.txt
absl-py==0.13.0
astor==0.8.1
attrs==19.3.0
backports.entry-points-selectable==1.1.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.24
distlib==0.3.2
dotmap==1.3.24
filelock==3.0.12
future==0.18.2
gast==0.2.2
gitdb==4.0.7
GitPython==3.1.18
glfw==2.1.0
google-pasta==0.2.0
grpcio==1.39.0
gym==0.17.1
h5py==3.3.0
imageio==2.9.0
importlib-metadata==4.6.3
inputs==0.5
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.37.0
lockfile==0.12.2
Markdown==3.3.4
matplotlib==3.4.2
mpi4py==3.0.3
numba==0.54.0
numpy==1.17.3
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.1
Pillow==8.2.0
platformdirs==2.2.0
protobuf==3.17.3
pybullet==3.1.7
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.1
pyzmq==22.2.1
quadprog==0.1.8
scipy==1.7.1
six==1.16.0
smmap==4.0.0
stable-baselines==2.10.0
tensorboard==1.15.0
tensorboardX==2.4
tensorflow==1.15.4
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.62.0
typing==3.7.4.1
typing-extensions==3.10.0.0
virtualenv==20.7.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
